requirements.txt
asgiref==3.7.2
Django==4.2.10
sqlparse==0.4.4
django-bootstrap-v5
django-debug-toolbar
psycopg==3.1.18
psycopg-binary==3.1.18
gunicorn
django-heroku
dj-database-url
python-decouple
social-auth-app-django
boto3
django-storages
django-multiupload~=0.6.1